✨️ (api): add error when the given
training id is not related to the given trigger training id

+++ api/src/devcomp/domain/usecases/delete-training-trigger.js
@@ -1,6 +1,23 @@
 import { withTransaction } from '../../../shared/domain/DomainTransaction.js';
+import { NotFoundError } from '../../../shared/domain/errors.js';
 
-const deleteTrainingTrigger = withTransaction(async function ({ trainingTriggerId, trainingTriggerRepository }) {
+const deleteTrainingTrigger = withTransaction(async function ({
+  trainingId,
+  trainingTriggerId,
+  trainingTriggerRepository,
+}) {
+  const trainingTriggers = await trainingTriggerRepository.findByTrainingId({
+    trainingId,
+  });
+
+  const isTrainingTriggerRelatedToTraining = trainingTriggers.some(
+    (trainingTrigger) => trainingTrigger.id === trainingTriggerId,
+  );
+
+  if (!isTrainingTriggerRelatedToTraining) {
+    throw new NotFoundError(`The training trigger ${trainingTriggerId} is not related to the training ${trainingId}`);
+  }
+
   return trainingTriggerRepository.deleteTrainingTrigger({
     trainingTriggerId,
   });
